refactor(gocoveragecheck): split coverage block diagnostics
